remove repeated loop
