+++ src/addtask.js
@@ -1,0 +1,4 @@
+
+export const read = document.querySelector('#new-task-input');
+export const addButton = document.querySelector('.add');
+
