+++ src/index.html
@@ -16,6 +16,7 @@
       src="https://kit.fontawesome.com/4aa116b8ee.js"
       crossorigin="anonymous"
     ></script>
+    <script src="./reload.js"></script>
     <title>ToDoList</title>
   </head>
   <body>
